pyproject.toml
[tool.poetry]
name = "django_sage_ticket"
version = "0.1.8"
description = "A Django-based ticketing system"
authors = ["Radin Ghaheremani <radin@sageteam.org>","Sepehr Akbarzadeh <sepehr@sageteam.org>"]
readme = "README.md"

packages = [
    { include = "sage_ticket" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = [
  { version = ">=4.2,<5.0", python = "3.10" },
  { version = ">=4.2,<5.3", python = ">=3.10" },      # Django 4.2 and 5.x for Python 3.10+
]
django-sage-tools = "^0.3.5"
django-modeltranslation = "^0.19.11"
django-import-export = "^4.3.3"
django-ckeditor-5 = "^0.2.15"
sorl-thumbnail = "^12.11.0"
readtime = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
tox = "^4.16.0"
coverage = "^7.6.0"
pre-commit = "^3.7.1"
sphinx = "^7.3.7"
sphinx-autobuild = "^2024.4.16"
sphinx-rtd-theme = "^2.0.0"
pylint = "^3.2.6"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.7"
commitizen = "^3.29.0"
tqdm = "^4.66.5"
mimesis = "^11.1.0"
pylint-django = "^2.5.5"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autobuild", "sphinx-rtd-theme"]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing --cov-report=html"
DJANGO_SETTINGS_MODULE = "ticketing.settings"
python_files = ["tests.py", "test_*.py"]
testpaths = ["tests"]
norecursedirs = [
    "migrations",
    "static",
    "media",
    "node_modules",
    "env",
    "venv",
    ".venv",
    "dist",
    "build"
]

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "ticketing/*",
    "*/apps.py",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.ruff]
line-length = 88
exclude = [
    "venv/*",
    ".venv/*",
    "build/*",
    "dist/*",
    ".git/*",
    "__pycache__/*",
    "*.egg-info/*",
    ".mypy_cache/*",
    ".pytest_cache/*",
    "migrations/*"
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
  | venv
  | env
  | __pycache__
  | static
  | media
  | node_modules
  | env
  | kernel
  | \.mypy_cache
  | \.pytest_cache
  | .*\.egg-info
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [
    "venv",
    ".venv",
    "build",
    "dist",
    ".git",
    "__pycache__",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
    "migrations",
    "static",
    "media",
    "node_modules",
    "env"
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.3"

[tool.commitizen.settings]
increment_types = ["feat", "fix"]

[tool.pylint]
disable = [
    "C0114",
    "C0115",
    "C0116",
    "E1101",
    "W0212",
    "C0330",
]
max-line-length = 88
ignore = [
    "migrations/*",
    "venv/*",
    "build/*",
    "dist/*",
    ".git/*",
    "__pycache__/*",
    "*.egg-info/*",
    ".mypy_cache/*",
    ".pytest_cache/*",
    "sage_ticket/ticket/tests/*",
]
load-plugins = [
    "pylint_django",
    "pylint.extensions.docparams",
]
django-settings-module = "kernel.settings"
good-names = [
    "qs",
    "pk",
    "id",
]
suggestion-mode = true
const-rgx = "([A-Z_][A-Z0-9_]*)|(__.*__)"
attr-rgx = "[a-z_][a-z0-9_]{2,30}$"
variable-rgx = "[a-z_][a-z0-9_]{2,30}$"
argument-rgx = "[a-z_][a-z0-9_]{2,30}$"
argument-name-hint = [
    "cls",
    "self",
]
method-rgx = "[a-z_][a-z0-9_]{2,30}$"
function-rgx = "[a-z_][a-z0-9_]{2,30}$"
class-rgx = "[A-Z_][a-zA-Z0-9]+$"
module-rgx = "(([a-z_][a-z0-9_]*)|(__.*__))$"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
